conntrack: Optimize recirculations.
In most cases, recirculations through conntrack can be much less costly.

If conntrack entry keys are changed outside of conntrack, the user must
call the pre-existing API, 'conntrack_clear()', which is modified in
this patch.

An implicit assumption is that a packet is processed to completion by a
PMD before returning to it's poll loop.

Signed-off-by: Darrell Ball <dlu998@gmail.com>